requirements.txt
Flask==2.3.2
gunicorn==20.1.0
sentence-transformers==2.2.2
torch>=1.12.0
Pillow==10.1.0
numpy==1.26.0
requests==2.31.0
waitress==3.0.2
